requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
Deprecated==1.2.14
distro==1.9.0
dnspython==2.4.1
exceptiongroup==1.2.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Limiter==3.5.0
flask-restx==1.0.6
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
limits==3.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mdurl==0.1.2
openai==1.9.0
ordered-set==4.1.0
packaging==23.2
pdf2image==1.17.0
pillow==10.2.0
pkgutil_resolve_name==1.3.10
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyJWT==2.6.0
pymongo==4.4.1
PyMySQL==1.0.2
pyrsistent==0.19.3
pytesseract==0.3.10
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
rich==13.7.0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.16.0
zipp==3.14.0
